Fix document sequence padding on PostgreSQL

+++ backend/alembic/sql/20260820_0001_canonical_v1.sql
@@ -13751,7 +13751,7 @@ BEGIN
     END IF;
     allocated_number := sequence.prefix || pg_catalog.lpad(
         sequence.next_value::text,
-        pg_catalog.greatest(sequence.padding,pg_catalog.length(sequence.next_value::text)),
+        pg_catalog.greatest(sequence.padding::integer,pg_catalog.length(sequence.next_value::text)),
         '0'
     ) || sequence.suffix;
     INSERT INTO "erp_core_commands"."command_scopes" VALUES
